CNA weightsheet recieves date via useLocation
