+++ src/components/dashboard/CNAdashboard.js
@@ -7,6 +7,15 @@ import { scale, list, message, alert } from "../../assets";
 import { fetchIt } from "../auth/fetchIt";
 import Footer from "../footer/Footer";
 
+const formattedDate = (date) => {
+  const myDate = date;
+  let year = myDate.toLocaleString("default", { year: "numeric" });
+  let month = myDate.toLocaleString("default", { month: "2-digit" });
+  let day = myDate.toLocaleString("default", { day: "2-digit" });
+  const formattedDate = year + "-" + month + "-" + day;
+  return formattedDate;
+};
+
 const CNAdashboard = () => {
   const navigate = useNavigate();
   const [name, setName] = useState("");
@@ -53,7 +62,7 @@ const CNAdashboard = () => {
         </h3>
       </div>
       <div className="grid lg:grid-cols-3 gap-10 lg:gap-0 justify-items-center lg:m-auto lg:w-2/3">
-        <Link to="/weeklysheet">
+        <Link to="/weeklysheet" state={{ date: "2023-04-08" }}>
           <div className=" rounded-lg p-6 shadow-lg shadow-sky-800/50 border-2 bg-amber-200/40 w-[300px] transform transition duration-500 hover:scale-90 hover:bg-sky-200/50">
             <h5 className="mb-5 text-lg font-body leading-tight  text-sky-800 font-semibold text-center">
               Weekly Weight Sheet
