+++ src/components/weightSheets/WeeklySheet.js
@@ -2,20 +2,12 @@ import React from "react"
 import NavBar from "../nav/NavBar"
 import { useState, useEffect } from "react"
 import { fetchIt } from "../auth/fetchIt"
-import { useNavigate } from "react-router-dom"
+import { useNavigate, useLocation } from "react-router-dom"
 import Swal from "sweetalert2";
 import withReactContent from "sweetalert2-react-content";
 
 
 
-const formattedDate = (date) => {
-  const myDate = date;
-  let year = myDate.toLocaleString("default", { year: "numeric" });
-  let month = myDate.toLocaleString("default", { month: "2-digit" });
-  let day = myDate.toLocaleString("default", { day: "2-digit" });
-  const formattedDate = year + "-" + month + "-" + day;
-  return formattedDate;
-};
 
 const formattedDateUI = (date) => {
   const myDate = date;
@@ -34,6 +26,8 @@ const WeeklySheet = () => {
     const [alerts, showAlerts] = useState(false)
     const [finalized, setFinalized] = useState(false)
     const navigate = useNavigate()
+    const location = useLocation();
+    const { date } = location.state;
     const MySwal = withReactContent(Swal);
 
     useEffect(() => {
@@ -48,9 +42,8 @@ const WeeklySheet = () => {
 
 
     useEffect(()=>{
-        const todaysDate = formattedDate(new Date())
         const API1 = "http://localhost:8000/weightsheets/create_all_weightsheets"
-        const API2 = `http://localhost:8000/weightsheets/detailedview_rd?date=${todaysDate}`
+        const API2 = `http://localhost:8000/weightsheets/detailedview_rd?date=${date}`
         const API3 ="http://localhost:8000/weights/closestdate_all?lookback=1week"
 
         //this function makes 3 api calls sequentially 
@@ -58,7 +51,7 @@ const WeeklySheet = () => {
             if (patientListWithWeights.length===0){
             await   fetchIt(API1, {
                         method: "POST",
-                        body: JSON.stringify({date: todaysDate}),
+                        body: JSON.stringify({date: date}),
                     })
                 }
             const weightsheets = await fetchIt(API2)
@@ -109,7 +102,7 @@ const WeeklySheet = () => {
             const address = `${API}/weights/${el.weight_id}`
             const requestBody = {
                 resident: el.resident_id,
-                date: formattedDate(new Date()),
+                date: date,
                 weight: el.weight,
             }
             promiseArray.push(
@@ -301,7 +294,7 @@ const WeeklySheet = () => {
         <div className="container mx-auto flex flex-col mt-20">
           <div className="flex md:justify-around justify-between sm:mx-10 mb-8 content-center items-center text-md sm:text-lg">
             <span>Weight Team Member: {employee.name}</span>
-            <span>Date: {formattedDateUI(new Date())}</span>
+            <span>Date: {date}</span>
             <button
               className={
                 "bg-sky-600 hover:bg-sky-500 uppercase text-sm py-2 px-6 mb-2 text-white font-bold rounded-full border border-blue focus:outline-none focus:border-sky-700 shadow-md "
